added stuffs

+++ download_paper.sh
@@ -2,7 +2,6 @@
 
 export START=$1
 export END=$2
-export SLEEP=0
 export DIRECTORY="paper"
 
 # create directory if does not exist
@@ -13,13 +12,13 @@ fi
 # create a function to make an NBER ID
 get_nber_id () {
     if (( $START < 10 )); then
-        NBER_ID="000${START}"
+        export NBER_ID="000${START}"
     elif (( $START >= 10 && $START < 100 )); then
-        NBER_ID="00${START}"
+        export NBER_ID="00${START}"
     elif (( $START >= 100 && $START < 1000 )); then
-        NBER_ID="0${START}"
+        export NBER_ID="0${START}"
     else
-        NBER_ID=${START}
+        export NBER_ID=${START}
     fi
 }
 
@@ -30,25 +29,20 @@ get_nber_id () {
 while (( $START < $END ))
 do
     get_nber_id
-    URL="https://www.nber.org/system/files/working_papers/w${NBER_ID}/w${NBER_ID}.pdf"
     if [ -e "${DIRECTORY}/${NBER_ID}.txt" ]
     then
-        echo "${DIRECTORY}/${NBER_ID}.txt already exists."
+        printf "[IGNORE 📁]: ${DIRECTORY}/${NBER_ID}.txt exists\n"
     else
-        wget $URL -O "${DIRECTORY}/${NBER_ID}.pdf"
-        STATUS=$?
-        if (( $STATUS == 0 ))
+        printf "[DOWNLOAD 💾]: ${DIRECTORY}/${NBER_ID}.pdf\n"
+        python3 src/download_paper.py
+        if [ -e "${DIRECTORY}/${NBER_ID}.pdf" ]
         then
-            printf "Download succeeded. Now let me sleep for ${SLEEP} seconds... \xF0\x9F\x98\xB4 \n"
+            pdftotext -layout "${DIRECTORY}/${NBER_ID}.pdf" "${DIRECTORY}/${NBER_ID}.txt"
+            rm -rf "${DIRECTORY}/${NBER_ID}.pdf"
+            printf "[SUCCEED ✅]: ${DIRECTORY}/${NBER_ID}.txt\n"
         else
-            printf "Download failed. Now let me sleep for ${SLEEP} seconds... \xF0\x9F\x98\xB4 \n"
+            printf "[404 📭]: ${DIRECTORY}/${NBER_ID}.pdf page not found\n"
         fi
-        sleep $SLEEP
-        pdftotext -layout "${DIRECTORY}/${NBER_ID}.pdf" "${DIRECTORY}/${NBER_ID}.txt"
-        rm -rf "${DIRECTORY}/${NBER_ID}.pdf"
     fi
     START=$(($START + 1))
 done
-
-# remove unwanted file
-rm -rf =
